Preparations for combining code
A method of obtaining what direction and how fast the paddles are moving when the ball collides with them has been added as a function in the paddle class: returnDy(). This is also a working point to return to if the combining of the code on my laptop and the code on my desktop does not go well.