Code combined but not working
The code has been updated with the offline code from my laptop but the ball does not move. This is because the trigonometry I use to determine how far the ball should move produces decimals while the x and y values of an object can only be updated by integers.