Testing with new angles
By increasing the speed of the ball, I find that the ball begins to move as decimals are automatically rounded. The trigonometry behind how the ball moves works properly, even if it is less accurate. This does cause problems with the ball clipping through the paddle and sometimes even the walls.