Code for changing angles not working
The new code added does not work. The ball instead becomes stuck in the paddle as the angle the ball is travelling at keeps changing. Try: looking at what I am changing the angle by (remember that the angle is in radians). Also try: Checking when the ball's x_direction & y_direction & angle are updated to see why it might be getting stuck.