Changes made to angles
By lowering the maximum angle possible, I have fixed a problem chich caused the ball to travel in a straight line vertically, making the ball impossible to hit. This error was caused by the calculated value for "dx" being so low that it was rounded to 0, meaning there was no horizontal movement.