Menu problem resolved
By reducing the clock speed of the menu to 10, it prevents too many keypresses being registered within a second, which prevents the issue of the selection going too fast. Also, by using "elif" instead of "if" for menus in the main loop, it stops the problem of the enter key remaining pressed.